Add a command line interface
Added a command line interface to last. It allows compile, run
or debug a last AST either using the parser or a JSON serlialized
version of the AST.

Includes a simple debugger that allows stepping displaying the
stack but not value access or break-points.

Running or debugging requires specifying a host .js file that
calls into the module.